diffraction: move ScanMaskEditor to its own module
A survey of what was left in bragg_peaks.py turned up something the earlier
cluster analysis had missed entirely: ScanMaskEditor is a *separate* 885-line
class with 46 methods, sharing the file with BraggPeaksPolymer for no reason. Its
only mention of BraggPeaksPolymer is a button tooltip string, and it uses none of
the module's helpers. So this is a file split rather than a refactor -- no
signature changes, no parameter threading, no shadowing risk.

bragg_peaks.py now contains exactly one class. 3663 lines, from 8206 when the
refactor started.

All three import paths still resolve to the same object, verified:
quantem.diffraction.ScanMaskEditor (what the tests use),
quantem.diffraction.bragg_peaks.ScanMaskEditor (preserved by the import that
edit_scan_mask needs anyway), and the new
quantem.diffraction.scan_mask_editor.ScanMaskEditor.

edit_scan_mask and create_interactive_circular_mask stay on the class: the editor
is constructed as ScanMaskEditor(self), so those are factories bound to the
analysis whose mask they edit, which is a legitimate reason to be a method.

Tests: 504 passed, 10 skipped. ScanMaskEditor already had coverage
(test_scan_mask_editor.py), so no new baseline tests were needed for this step.
